finished most of user auth
