+++ package.json
@@ -41,6 +41,7 @@
     "prop-types": "^15.7.2",
     "react": "^16.4",
     "react-dom": "^16.4",
+    "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "webpack": "^5.38.1",
     "webpack-cli": "^4.2.0"
