+++ server/lib/middleware/userInViews.js
@@ -1,0 +1,4 @@
+module.exports = () => (req, res, next) => {
+  res.locals.user = req.user;
+  next();
+};
