+++ server/routes/home.js
@@ -1,0 +1,10 @@
+const express = require('express');
+
+const router = express.Router();
+
+/* GET home page. */
+router.get('/', (req, res) => {
+  res.render('index', { title: 'Auth0 Webapp sample Nodejs' });
+});
+
+module.exports = router;
