+++ server/routes/index.js
@@ -1,7 +1,15 @@
 const events = require('./events');
 const hotels = require('./hotels');
+const home = require('./home');
+const users = require('./users');
+const auth = require('./auth');
+const flights = require('./flights');
 
 module.exports = {
   events,
   hotels,
+  home,
+  users,
+  auth,
+  flights,
 };
